requirements.txt
discord.py
yt-dlp
PyNaCl
static-ffmpeg
configparser
asyncio